+++ src/api/Domain/Controllers/Api/Sponsors/BeneficiaryIdReference/Get/Models/DonationsResponse.cs
@@ -3,14 +3,14 @@
     public class DonationsResponse
     {
         public DonationsResponse(
-            int totalInHundreds,
-            int monthlyInHundreds)
+            long totalInHundreds,
+            long monthlyInHundreds)
         {
             TotalInHundreds = totalInHundreds;
             MonthlyInHundreds = monthlyInHundreds;
         }
 
-        public int TotalInHundreds { get; }
-        public int MonthlyInHundreds { get; }
+        public long TotalInHundreds { get; }
+        public long MonthlyInHundreds { get; }
     }
 }
